🚨 Fix lint warnings
